Add files via upload
Project  is about Hall Management System in a University , a website. It deals with the hall dues of the students , Meal on/off,controlling system and notice board also. Each Students easily get message or informations throught this easily.It is PHP project .Also accompanies with  MySQL,HTML,CSS,BOOTSTRAP ,AJAX and XML